rename denyList.ts as deny-list.ts
